requirements.txt
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
wheel==0.45.1
